Master (#19)
* ci: add .deepsource.toml

* refactor: remove unnecessary f-string (#16)

It is wasteful to use `f-string` mechanism if there are no expressions to be extrapolated. It is recommended to use regular strings instead.

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>

* refactor: convert string with anomalous backslash into a raw string (#18)

Backslash is present in the literal string but is not a valid escape sequence. If it is intended to be an escape sequence, use the correct escape characters. If it is intended to be a literal backslash, it can either be replaced with with an escaped backslash `\\`, or you can add an `r` prefix to make it a "raw" string.

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>

---------

Signed-off-by: gitworkflows <118260833+gitworkflows@users.noreply.github.com>
Co-authored-by: deepsource-io[bot] <42547082+deepsource-io[bot]@users.noreply.github.com>
Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>